requirements.txt
# Core
Django>=4.0,<=4.1
psycopg2-binary>=2.8
djangorestframework==3.14.0
drf-yasg==1.21.5
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
Pillow>=9.3
celery==5.2.7
requests==2.28.2
paho-mqtt==1.6.1

# Test
faker==16.8.1
ipython==7.21.0
